Move RepoMan Config Page to Config-Area
