+++ Libs/libMacGitverCore/Config/Ui/RepoManConfigPage.ui
